chore(review): move isLocalLink to correct utils folder
