+++ src/shared/utils/link.ts
